Create README.md
Loan Approval dataset is a popular Kaggle dataset for whether loan would be approved or not based on the loan details such as the loan amount, loan term etc. as well as the loan borrower's detail such as annual income, number of dependents etc. This Model works on the above mentioned dataset and predicts whether or not the loan would be approved.